+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.9.4
+## 0.9.5
 - Fix issue [#30](https://github.com/RafaelBarbosatec/qr_code_dart_scan/issues/30). Adds `forceReadPortrait` in `QRCodeDartScanView` to force read portrait image.
+
+## 0.9.4
 - adds methods `toggleFlash`,`setFlash`,`setFlashAuto` in QRCodeDartScanController.
 
 ## 0.9.3